requirements.txt
Flask==2.3.0
requests==2.28.1
python-dotenv==0.21.0
